+++ internal/extproc/translator/gemini_helper.go
@@ -11,10 +11,7 @@ import (
 	"mime"
 	"net/url"
 	"path"
-	"strconv"
 
-	"github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
-	"github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
 	"github.com/google/uuid"
 	"google.golang.org/genai"
 
@@ -22,7 +19,9 @@ import (
 )
 
 const (
-	GCPModelPublisherGoogle    = "google"
+	// GCPModelPublisherGoogle is using by Gemini model family.
+	GCPModelPublisherGoogle = "google"
+	// GCPModelPublisherAnthropic is using by Calude model family
 	GCPModelPublisherAnthropic = "anthropic"
 	GCPMethodGenerateContent   = "generateContent"
 )
@@ -494,33 +493,4 @@ func toLogprobs(logprobsResult genai.LogprobsResult) openai.ChatCompletionChoice
 func buildGCPModelPathSuffix(publisher, model, gcpMethod string) string {
 	pathSuffix := fmt.Sprintf("publishers/%s/models/%s:%s", publisher, model, gcpMethod)
 	return pathSuffix
-}
-
-// buildGCPRequestMutations creates header and body mutations for GCP requests
-// It sets the ":path" header, the "content-length" header and the request body.
-func buildGCPRequestMutations(path string, reqBody []byte) (*ext_procv3.HeaderMutation, *ext_procv3.BodyMutation) {
-	// Create header mutation
-	headerMutation := &ext_procv3.HeaderMutation{
-		SetHeaders: []*corev3.HeaderValueOption{
-			{
-				Header: &corev3.HeaderValue{
-					Key:      ":path",
-					RawValue: []byte(path),
-				},
-			},
-			{
-				Header: &corev3.HeaderValue{
-					Key:      "content-length",
-					RawValue: []byte(strconv.Itoa(len(reqBody))),
-				},
-			},
-		},
-	}
-
-	// Create body mutation
-	bodyMutation := &ext_procv3.BodyMutation{
-		Mutation: &ext_procv3.BodyMutation_Body{Body: reqBody},
-	}
-
-	return headerMutation, bodyMutation
 }
